Pass handler to tar_file_iterator
The exception handler is currently away from the `tar_file_iterator`. I am not sure that it is intentional or just a typo. 

I feel that it is a typo since the handler is also used to handle the exceptions caused by a sample (instead of a tar file) in group_by_keys. Thus I proposed this one-line change to resolve this issue.