Added link to WebDataset Format Specification.
The README.md and readme.ipynb files have been updated to include a new
link to the WebDataset Format Specification document. This provides
users with a resource for a more detailed and longer specification
of the WebDataset format. The update is consistent across both the
Markdown and Jupyter Notebook documentation files.