Updated dependencies and warning message in WebDataset
A column-store example has been added.
In the requirements.dev.txt, several Python packages related to
PyTorch and Ray have been commented out since they don't work in 24.04;
this needs to be fixed.
The the WebDataset class has been updated to correct the class name
in the message when shardshuffle is set to None.