WebDataset Library Updates
Several updates have been made to the `WebDataset` library to improve
warning messages and functionality. In `webdataset/compat.py`, the
`WebDataset` class now issues a warning when `shardshuffle` is set
for resampled datasets, and another warning when `shardshuffle` is
set to `True`, advising to use a positive integer, `0`, or `False`
instead. Additionally, the default value for `shardshuffle` is set to
`100` when it is `True`.

In `webdataset/pipeline.py`, the `DataPipeline` class's `with_length`
method has been updated to accept a `silent` parameter. When `silent`
is `False`, a warning is issued to clarify that the `with_length`
method only sets the value of `__len__` for compatibility with certain
training environments and does not affect the actual number of samples.

Lastly, in `webdataset/utils.py`, the check for whether an object
is an `Iterator` has been removed from the `is_iterable` function,
leaving the check for `Iterable` to determine if an object is iterable.