Update Documentation and Dependencies, Refactor Imports
Updated the `mkdocs.yml` configuration to include a new
navigation structure and theme settings. Added several plugins
to the documentation build process, including `mkdocstrings` and
`mkdocs-material`. The `requirements.dev.txt` was updated to include
new packages for documentation generation. Refactored the `tasks.py`
script to include a new `mkdocs` task for building documentation and
updated the `gendocs` task to handle Jupyter notebooks conversion. In
`webdataset/__init__.py`, the import statement was modified to include
`FluidInterface`. The import of `WebDataset` in `wids/wids_bench.py`
was changed to use the `webdataset` module directly.

  docsrc/webdataset.md: It manages WebDataset operations for data processing and I/O.
  docsrc/wids.md: It manages data sampling in distributed machine learning.