Update wids to handle local tars efficiently without copying them to cache
wids handles local datasets, but it copies them to the /tmp folder, which can be inefficient if the dataset is already stored on your machine. To avoid unnecessary copying, we first check whether the dataset exists locally. If it does, we open it directly, otherwise, it is downloaded and saved to the /tmp folder.